refactor(geode-core): make assembly directory-backed with p1/nedelec/fe/sparse/surface submodules
Convert the single-file `assembly` module into a directory-backed group
and fold the four sibling global-assembly modules under it:

- assembly.rs        -> assembly::p1      (GlobalSystem, assemble_global_p1, ...)
- nedelec_assembly.rs-> assembly::nedelec (assemble_global_nedelec* family, PML builders, scatter maps)
- fe_assemble.rs     -> assembly::fe      (fe_assemble, DirichletBc, ElementType)
- sparse.rs          -> assembly::sparse  (SparseSystem, global_system_to_sparse)
- silvermuller.rs    -> assembly::surface (Silver-Muller / surface-mass assembly)

New assembly/mod.rs carries group rustdoc and re-declares the five
submodules. lib.rs drops the four removed `pub mod` decls (keeping
`pub mod assembly`) and converts the five flat-root re-export blocks into
`#[deprecated]` shims pointing at the canonical `assembly::{p1,nedelec,fe,
sparse,surface}` paths. The prelude gains the high-traffic entry points
from the canonical paths.

All in-crate references to the removed module paths and relocated
`assembly` items are migrated in this PR (lib code, #[cfg(test)] code, and
intra-doc links), since the old module paths disappear entirely. Pure
refactor with `git mv` to preserve history; no behavior change.

Closes #380. Part of epic #377.
