+++ crates/geode-core/src/assembly/fe.rs
@@ -39,9 +39,9 @@ use burn::tensor::backend::Backend;
 use faer::Mat;
 
 use crate::TetMesh;
-use crate::assembly::{assemble_global_p1, upload_mesh};
+use crate::assembly::nedelec::assemble_global_nedelec_with_epsilon;
+use crate::assembly::p1::{assemble_global_p1, upload_mesh};
 use crate::eigen::{EigenError, apply_dirichlet_bc, burn_matrix_to_faer};
-use crate::nedelec_assembly::assemble_global_nedelec_with_epsilon;
 
 /// Selector for the finite element type passed to [`fe_assemble`].
 ///
