+++ crates/geode-core/src/assembly/sparse.rs
@@ -1,9 +1,9 @@
 //! CSR/CSC projection of the assembled dense `GlobalSystem` into faer's
 //! sparse representation, with optional Dirichlet boundary reduction.
 //!
-//! The dense `K` and `M` that come off [`crate::assembly::assemble_global_p1`]
+//! The dense `K` and `M` that come off [`crate::assembly::p1::assemble_global_p1`]
 //! are full `[n_dof, n_dof]` Burn tensors but the underlying P1 stencil is
-//! sparse — every entry not in the assembled [`crate::assembly::SparsityPattern`] is exactly
+//! sparse — every entry not in the assembled [`crate::assembly::p1::SparsityPattern`] is exactly
 //! zero. We:
 //!
 //! 1. Pull the dense matrices to host f64 once (`burn_matrix_to_faer`).
@@ -25,7 +25,7 @@
 use burn::tensor::backend::Backend;
 use faer::sparse::{SparseColMat, Triplet};
 
-use crate::assembly::GlobalSystem;
+use crate::assembly::p1::GlobalSystem;
 use crate::eigen::burn_matrix_to_faer;
 
 /// Errors produced by the dense → sparse projection.
